requirements.txt
transformers
torch
wordcloud
matplotlib
nltk
numpy
pillow
tqdm